separate data to make tests work properly
