+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "scripts": {
     "start": "node server",
+    "test": "jest -i ./__test__",
     "server": "nodemon server",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
@@ -29,10 +30,14 @@
     "passport-jwt": "^4.0.0"
   },
   "devDependencies": {
+    "@types/jest": "^24.0.13",
     "concurrently": "^4.1.0",
     "dotenv": "^8.0.0",
+    "faker": "^4.1.0",
+    "jest": "^24.8.0",
     "nodemailer": "^6.2.1",
-    "nodemon": "^1.19.0"
+    "nodemon": "^1.19.0",
+    "supertest": "^4.0.2"
   },
   "repository": {
     "type": "git",
